+++ android/src/main/java/com/dynamsoft/flutter_barcode_sdk/FlutterBarcodeSdkPlugin.java
@@ -94,6 +94,17 @@ public class FlutterBarcodeSdkPlugin implements FlutterPlugin, MethodCallHandler
                 int ret = mBarcodeManager.setBarcodeFormats(formats);
                 result.success(ret);
             }
+            break;
+            case "getParameters": {
+                result.success(mBarcodeManager.getParameters());
+            }
+            break;
+            case "setParameters": {
+                final String params = call.argument("params");
+                int ret = mBarcodeManager.setParameters(params);
+                result.success(ret);
+            }
+            break;
             default:
                 result.notImplemented();
         }