+++ example/lib/web.dart
@@ -1,3 +1,5 @@
+import 'dart:convert';
+
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
 import 'package:flutter_barcode_sdk/dynamsoft_barcode.dart';
@@ -27,6 +29,19 @@ class _WebState extends State<Web> {
   Future<void> initBarcodeSDK() async {
     _barcodeReader = FlutterBarcodeSdk();
     await _barcodeReader.setBarcodeFormats(BarcodeFormat.ALL);
+    // Get all current parameters.
+    // Refer to: https://www.dynamsoft.com/barcode-reader/parameters/reference/image-parameter/?ver=latest
+    String params = await _barcodeReader.getParameters();
+    // Convert parameters to a JSON object.
+    dynamic obj = json.decode(params);
+    // Modify parameters.
+    if (obj['ImageParameter'] != null) {
+      obj['ImageParameter']['DeblurLevel'] = 5;
+    } else
+      obj['deblurLevel'] = 5;
+    // Update the parameters.
+    int ret = await _barcodeReader.setParameters(json.encode(obj));
+    print('Parameter update: $ret');
   }
 
   // Platform messages are asynchronous, so we initialize in an async method.