+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## 1.2.0
+* Added two methods: getParameters() and setParameters().
+
+## 1.1.1
+* Fixed the crash issue caused by building release app for Android.
+
 ## 1.1.0
 * Added setBarcodeFormats() method.